backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-filter==24.2
django-cors-headers==4.3.1

djangorestframework-simplejwt==5.4.0
drf-yasg==1.21.8
drf-spectacular==0.29.0

psycopg2-binary==2.9.10


celery==5.3.1
django-celery-beat==2.7.0   # supports Django 5 and Python 3.11+

django-storages==1.14.2
boto3==1.34.141

python-dotenv==1.0.1
django-environ==0.11.2

pytest==8.2.0
pytest-django==4.8.0
pytest-cov==5.0.0
coverage==7.6.0
factory-boy==3.3.0
